♻️ make actUponStore a member of Internal
