+++ packages/atom.io/internal/src/transaction/act-upon-store.ts
@@ -1,0 +1,19 @@
+import type { TransactionToken, ƒn } from "atom.io"
+
+import { NotFoundError } from "../not-found-error"
+import type { Store } from "../store"
+import { withdraw } from "../store"
+
+export function actUponStore<ƒ extends ƒn>(
+	token: TransactionToken<ƒ>,
+	id: string,
+	store: Store,
+): (...parameters: Parameters<ƒ>) => ReturnType<ƒ> {
+	return (...parameters: Parameters<ƒ>): ReturnType<ƒ> => {
+		const tx = withdraw(token, store)
+		if (tx) {
+			return tx.run(parameters, id)
+		}
+		throw new NotFoundError(token, store)
+	}
+}
