+++ packages/atom.io/realtime-server/src/realtime-action-synchronizer.ts
@@ -1,6 +1,7 @@
-import * as AtomIO from "atom.io"
+import type * as AtomIO from "atom.io"
 import {
 	IMPLICIT,
+	actUponStore,
 	findInStore,
 	getFromStore,
 	setIntoStore,
@@ -52,7 +53,7 @@ export function realtimeActionSynchronizer({
 			const performanceKeyStart = `${performanceKey}:start`
 			const performanceKeyEnd = `${performanceKey}:end`
 			performance.mark(performanceKeyStart)
-			AtomIO.actUponStore<ƒ>(tx, update.id, store)(...update.params)
+			actUponStore<ƒ>(tx, update.id, store)(...update.params)
 			performance.mark(performanceKeyEnd)
 			const metric = performance.measure(
 				performanceKey,
